requirements.txt
botocore==1.15.49
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
DateTime==4.3
docutils==0.15.2
Flask==1.1.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
pathspec==0.5.9
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
requests==2.20.1
scikit-learn==0.23.1
scipy==1.5.1
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
termcolor==1.1.0
threadpoolctl==2.1.0
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==1.0.1
wincertstore==0.2
zope.interface==5.1.0
